Initial PyPlant project
This project is intended to parse python code and analyse it to generate
classdiagramms (UML). The first primary usage will be to generate
plantuml to generate the diagramm format (therefore the name), but the
project is planned to be easy extendable to generate output for other
UML generators. It is also planned to be able to design it so it can be
extended with parsers for different languages then python.